requirements.txt
# Bio-Medical AI Competition - Simple Starter Kit

# Core dependencies
transformers>=4.21.0
torch>=1.12.0
tqdm>=4.64.0
pandas>=1.3.0

# Optional dependencies for different model types
openai>=1.0.0  # For ChatGPT models
google-generativeai>=0.3.0  # For Gemini models (optional)
google-cloud-aiplatform>=1.38.0  # For Vertex AI/Gemini with service account auth

# Utility packages
numpy>=1.21.0
scikit-learn>=1.0.0  # For additional metrics (optional)
requests>=2.31.0

# Development dependencies (optional)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Tool call
tooluniverse>=1.0.7
python-dotenv>=0.19.0  # For environment variable management
openai-harmony>=1.0.0  # For GPT-OSS-20B support